Replay HTML and htmlwidgets as iframe
Currently when HTML and htmlwidgets are called in RStudio, they automatically render in RStudio viewer and fail to be captured by `evaluate`.

This bypasses the `print` for these two classes and instead save them to a local html file. The html is then embedded as an iframe in the reference.