Rewrite example rendering
First step is to convert mixed R-Rd to R, translating (e.g.)

\dontest{
  1 + 2
}

to

  1 + 2

This is suboptimal because it leaves evidence of internal Rd tags that users probably don't care about, but there's no other way to get fidelity whitespace treatment without competely rewriting the parser.

The key problem is that in (e.g)

if (TRUE) {
  \dontrun{
    "Hello"
  }
}

the whitespace before `\d` and after the terminal `}` does no belong to the tag_dontrun object, so a pure recursive parser has no way to handle it.

Evaluating all the code at once makes the topic image paths simpler, making it possible to remove the R6 dependency.

Fixes #1087