Alert user that `build_home()` does not update images on home page
Currently the user needs to run `build_reference()` in order to update the images on the home page, see #1225. This patch adds an explanation of this to the documentation. An alternative would be to let `build_home()` do the updating of the images itself. A pull request for that was however rejected by Hadley, see #1226