Explicitly set the remote's branch before fetching
Travis-CI uses a shallow clone of the repository, which means that the
'origin' remote only tracks the HEAD branch. Even if we later fetch a
different branch the worktree command will still fail, because the
branch is not tracked by the remote. Using `git set-branches` we can
explicitly add the branch we are trying to deploy to be tracked by the
remote, which fixes the issue.

Fixes #1271