Fix Typo - Missing ":" - Line 167:17 - build.R
I was testing out the example displayed in section "YAML config - navbar" in build_site() reference page (https://pkgdown.r-lib.org/reference/build_site.html). I got an error when I blindly copied the example and executed it on my computer. Later I found the cause to be a missing colon in one key-value pair, which is in line 167 column 17 in https://github.com/r-lib/pkgdown/edit/master/R/build.r .

I realized that its just a typo on the webpage, rather than a bug in code - perhaps that's why none of the GitHub actions are able to detect this typo.