Anchor & news tweaks (#1830)
This PR started out as an attempt to fix scrollspy on news, but rapidly spiralled out of control. 

It now includes overall improvements to the news page (bringing the headers back closer to default styling, using semantic html for CRAN release date, and only show date if found). Fixing the anchors on the news page revealed that life was unnecessarily complicated because each section was given two ids,  one by pandoc on the `div` and one by the js TOC on the heading, so I now unify them by moving the section id to the first heading. I also realised that because ids are now made unique, I could remove the pseudo-namespacing done in some of the templates.

The change to the location of the ids broke the tabset code, so I rewrote to work with xml nodes rather than string ids. Looking closely at this code made me realise the ids for the aria labelling weren't set up quite right.

Along the way, I discovered that the scrollspy and headroom were set up using both explicit js and body attributes, so I standardised on the js.

Fixes #1821