Remove unneeded pkg argument from markdown functions
This mostly seems to be some remnant of earlier usage. The only real place it was needed which `markdown_text_inline()`, which now takes a `where` argument which will typically be generated by `pkgdown_field()`. I also tweaked its error to be slightly easier to understand (IMO).

Extracted out of #1967