Make the navbar a proper menu component
Fixes #2320

I tested the visual appearance manually in a few places, and it looks good to me. @maelle it'd be useful if you could kick the tires on it a bit too.

While working on this code, I realised that we'd been incorrectly labelling the search box as "toggle navigation" to screen readers :facepalm: I've fixed that in this PR too.