fix: improve support for math rendering with CDN (#2833)
This PR includes katex and mathjax dependencies via the head template instead of via `cached_dependency()`.

This simplifies the katex loading, which otherwise would require caching ~60 katex files. 

But once katex material moved to the head template, then it doesn't make sense to have a separate approach for mathjax, so that is moved to the head template too.

Closes #2704