Improve CSS selection of rendered images (#2936)
Closes #2920

The main change is the addition of a CSS class for images rendered by knitr, via setting `fig.class` to `r-plt` (the same class used by downlit) in a knitr hook.

Rendered images in examples and in quarto vignettes already had unique CSS classes, so those were added to the selector.